requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
scikit-learn
openai
python-dotenv